requirements.txt
ultralytics
opencv-python
websockets
geocoder
pynmea2
pyserial
numpy
torch
torchvision
